+++ package.json
@@ -54,6 +54,7 @@
     "redux": "^3.7.1",
     "redux-actions": "^2.2.1",
     "redux-logger": "^3.0.6",
+    "redux-mock-store": "^1.3.0",
     "redux-thunk": "^2.2.0",
     "redux-undo": "^0.6.1",
     "reselect": "^3.0.1",
