+++ src/actions/todo/index.js
@@ -1,9 +1,12 @@
 let nextTodoId = 0;
-export const addTodo = text => ({
-  type: 'ADD_TODO',
-  id: nextTodoId++,
-  text,
-});
+export const addTodo = text => {
+  nextTodoId += 1;
+  return {
+    type: 'ADD_TODO',
+    id: nextTodoId,
+    text,
+  };
+};
 
 export const setVisibilityFilter = filter => ({
   type: 'SET_VISIBILITY_FILTER',
