+++ src/components/ToDo/Link.js
@@ -1,25 +1,29 @@
-import React from 'react'
-import PropTypes from 'prop-types'
+import React from 'react';
+import PropTypes from 'prop-types';
 
 const Link = ({ active, children, onClick }) => {
   if (active) {
-    return <span>{children}</span>
+    return <span>{children}</span>;
   }
 
   return (
-    <a href="#" onClick={e => { // eslint-disable-line jsx-a11y/href-no-hash
-      e.preventDefault()
-      onClick()
-    }}>
+    // eslint-disable-next-line
+    <a
+      href="#"
+      onClick={e => {
+        e.preventDefault();
+        onClick();
+      }}
+    >
       {children}
     </a>
-  )
-}
+  );
+};
 
 Link.propTypes = {
   active: PropTypes.bool.isRequired,
   children: PropTypes.node.isRequired,
-  onClick: PropTypes.func.isRequired
-}
+  onClick: PropTypes.func.isRequired,
+};
 
-export default Link
+export default Link;
