+++ src/selectors/todo/index.js
@@ -1,0 +1,30 @@
+import { createSelector } from 'reselect';
+
+const getVisibilityFilter = state => state.visibilityFilter;
+const getTodos = state => state.todos.present; // !!!!used redux-undo
+
+export const getVisibleTodos = createSelector(
+  [getVisibilityFilter, getTodos],
+  (visibilityFilter, todos) => {
+    switch (visibilityFilter) {
+      case 'SHOW_COMPLETED':
+        return todos.filter(todo => todo.completed);
+      case 'SHOW_ACTIVE':
+        return todos.filter(todo => !todo.completed);
+      default:
+        return todos;
+    }
+  },
+);
+
+export const makeGetVisibleTodos = () =>
+  createSelector([getVisibilityFilter, getTodos], (visibilityFilter, todos) => {
+    switch (visibilityFilter) {
+      case 'SHOW_COMPLETED':
+        return todos.filter(todo => todo.completed);
+      case 'SHOW_ACTIVE':
+        return todos.filter(todo => !todo.completed);
+      default:
+        return todos;
+    }
+  });
